Allow webhooks for Stripe Connect
Stripe uses the user_id parameter when the webhook if for Connect. This
commit extracts user_id from the passed in params and sets it at the
root level of the json. This should be safe enough as there is no other
use of user_id in the stripe api.