Add StripeMock.mock method
This method takes a block, and yields it between StripeMock.start and
StripeMock.stop. This ensures we don't forget to call StripeMock.stop at
the appropriate time.

For example:

test ".stripe_customer fetches the account's Stripe::Customer from Stripe and returns it" do
  StripeMock.mock do
    customer = account.stripe_customer
    assert_equal Stripe::Customer, customer.class
  end
end

We use `begin ... ensure` to ensure that StripeMock.stop is called, even
if the block raises an exception.