Allow overwriting of the default currency
In my application, when we create customers, we aren't specifying the
currency. Stripe's documentation for "Create a customer" doesn't state
that you can specify a currency, so I'm guessing that it is somehow
inferred.

As my application does not use USD, the check for matching currencies
that was added in 0bb51d14e3668d4b026bd4de6c810a402ab90252 broke my
tests.

Allowing to specify a default currency other than usd seemed like the
most straightforward approach to get this working.

I couldn't figure out a good approach to get it to work for the server
spec, as I can't easily overwrite the default for the server (as the
server is started once before all the specs). It should still work
though even if you are using the server. Due to my change having little
chance of introducing a regression, I decided to only make a spec for
the instance version.