Fix how nested data is updated on customers
Stripe interprets empty strings as "delete this value". This has several implications for how parameters are treated.

Simply merging the new params with the old customer-object overwrites any current customer attribute containing a hash, as all nested customer attributes are included in the params, regardless of if changes have been made.

This issue has been partially addressed in previous commits, but it does not account for all nested attributes, such as `metadata` and `shipping`. This code has been refactored to work with any nested attribute.

Additionally, passing an empty string as a parameter now sets the value to `nil`, which is closer to how the Stripe API works.