Fix List initialize error with deleted records
When deleting records, the record is marked internally as deleted but
other data fields are cleared, including `created`. The deleted record
is passed to StripeMock::Data::List.new from their respective #list
methods, e.g. Stripe::Customer.list. The sorting of these records in
initialize raised an ArgumentError (comparison of Integer with nil
failed).

This fix addresses the issue by omitting deleted records from the data
as that seems to be the more intuitive behavior. Minor refactoring was
also performed to normalize the sorting of data whether it is in Hash
or StripeObject form.