Add `require` option to require additional file
When booting the stripe-mock-server it can be useful to have Ruby
`require` in additional files if you need to patch `StripeMock` in some
way. This is mostly useful for differences when you're running on an
older version of the Stripe API and need to tweak `StripeMock`'s
behavior until your Stripe API version is upgraded to be compatible.

When using `StripeMock` directly within a Rails app, for example, this
is unneeded as you can do the same thing via a Rails initializer file.
But if running the `StripeMock` server standalone - for use with front
end tests, for example, you might pass this new flag like this:

--require config/initializers/stripe_ruby_mock