Add drb gem to Gemspec
In Ruby 3.4, several [gems are removed from the default gem
list](https://stdgems.org/new-in/3.4/).

This means that calling "require 'gem'" will fail if they aren't
explicitly installed via dependencies or a Gemfile.

This project calls require on 'drb' which is one of these gems. That
means anyone using stripe-ruby-mock with Ruby 3.3 will see a warning,
and with 3.4 an error.

This PR adds 'drb' to the dependencies in the gemspec. I chose to use
the latest version with flexibility on the patch release, but let me
know if another specification could be better.